requirements.txt
anytree
jupyterlab
imblearn
lime
matplotlib==3.4.1
minds-kit
namedlist
nltk
notebook
numpy
opencv-python
pandas==1.3.4
Pillow
pysmt==0.9.0
python-sat==0.1.8.dev1
scikit-learn==1.0.2
scipy==1.9.1
shap
tensorflow==2.13.0
torch 
torchvision
tqdm
wittgenstein
xgboost==1.7.3
